requirements.txt
# General dependencies for data handling
numpy==1.25.2
pandas==2.1.1
scikit-learn==1.3.0

# For handling imbalanced data (SMOTE)
imbalanced-learn>=0.10.1

# Image processing
Pillow==10.0.1
opencv-python==4.8.1.78
torchvision==0.16.0  # Optional for additional image augmentations using PyTorch

# Text processing
nltk==3.9.1  # Natural Language Toolkit for text augmentation
transformers==4.34.0  # For back-translation using Hugging Face transformers

# FastAPI and API dependencies
fastapi==0.103.1  # Web framework for building APIs
uvicorn==0.24.0  # ASGI server for running FastAPI
python-dotenv==1.0.0  # For managing environment variables (.env)

# Testing
pytest==7.4.2  # Testing framework
pytest-cov>=4.2.0  # Code coverage reporting with pytest

# Logging and monitoring
loguru==0.7.0  # Optional: alternative to Python's logging module for better logging experience

gunicorn==21.2.0  # Production WSGI server

celery>=5.4.0  # Task queue for background processing
redis>=5.1.1  # Redis as the broker for Celery tasks
